fix(engine/mysql): Support nullable fields in joins for MySQL engine (#1249)
* Fix nullable fields in joins for MySQL engine

+++ internal/endtoend/testdata/join_right/mysql/go/models.go
@@ -7,10 +7,10 @@ import (
 )
 
 type Bar struct {
-	ID int32
+	ID int64
 }
 
 type Foo struct {
-	ID    int32
+	ID    int64
 	BarID sql.NullInt32
 }
